+++ test/framework/src/main/java/org/opensearch/test/telemetry/MockTelemetry.java
@@ -11,8 +11,10 @@ package org.opensearch.test.telemetry;
 import org.opensearch.telemetry.Telemetry;
 import org.opensearch.telemetry.TelemetrySettings;
 import org.opensearch.telemetry.metrics.Counter;
+import org.opensearch.telemetry.metrics.Histogram;
 import org.opensearch.telemetry.metrics.MetricsTelemetry;
 import org.opensearch.telemetry.metrics.noop.NoopCounter;
+import org.opensearch.telemetry.metrics.noop.NoopHistogram;
 import org.opensearch.telemetry.tracing.TracingTelemetry;
 import org.opensearch.test.telemetry.tracing.MockTracingTelemetry;
 
@@ -44,6 +46,11 @@ public class MockTelemetry implements Telemetry {
             @Override
             public Counter createUpDownCounter(String name, String description, String unit) {
                 return NoopCounter.INSTANCE;
+            }
+
+            @Override
+            public Histogram createHistogram(String name, String description, String unit) {
+                return NoopHistogram.INSTANCE;
             }
 
             @Override